Move CreatePaypalPaymentData into readerRevenueApis
