+++ assets/helpers/paymentIntegrations/payPalPaymentAPICheckout.js
@@ -12,23 +12,9 @@ import type { IsoCountry } from 'helpers/internationalisation/country';
 import type { ReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
 import type { Participations } from 'helpers/abTests/abtest';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
-import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import type { OptimizeExperiments } from 'helpers/tracking/optimize';
+import { type CreatePaypalPaymentData } from 'helpers/paymentIntegrations/readerRevenueApis';
 
-// ----- Types ----- //
-
-// Data that should be posted to the payment API to get a url for the PayPal UI
-// where the user is redirected to so that they can authorize the payment.
-// https://github.com/guardian/payment-api/blob/master/src/main/scala/model/paypal/PaypalPaymentData.scala#L74
-export type CreatePaypalPaymentData = {|
-  currency: IsoCurrency,
-  amount: number,
-  // Specifies the url that PayPal should make a GET request to, should the user authorize the payment.
-  // Path of url should be /paypal/rest/return (see routes file)
-  returnURL: string,
-  // Specifies the url that PayPal should make a GET request to, should the user not authorize the payment.
-  cancelURL: string,
-|}
 
 // ----- Functions ----- //
